trading_dev/requirements.txt
flask==3.0.3
gunicorn==23.0.0
pandas==2.2.3
portalocker==2.10.1
psycopg2-binary==2.9.10
pytest-asyncio==0.24.0
pytest-cov==4.1.0
pytest-mock==3.14.0
pytest==8.2.2
pytz==2024.2
pyyaml==6.0.2
sqlalchemy==2.0.40
telethon==1.37.0
apscheduler==3.10.4
mysql-connector-python==9.1.0
